requirements.txt
streamlit
python-dotenv
groq
faiss-cpu
chromadb
PyMuPDF
pandas
